feat: :sparkles: I created a SmartRoute component for managing authentication routes.
I created a SmartRoute component that checks for the presence of a token cookie.
If authentication is required and the token is missing, it redirects the user to the login page.
Conversely, if authentication is not required and the token is present, it redirects the user to the home page.
I then used SmartRoute in the Router component to protect the AdminPage and set up the login page.